+++ Downloader.Business.Contract/IDownloadBuilder.cs
@@ -8,6 +8,6 @@ namespace Toqe.Downloader.Business.Contract
 {
     public interface IDownloadBuilder
     {
-        IDownload Build(Uri url, int bufferSize, int? offset, int? maxReadBytes);
+        IDownload Build(Uri url, int bufferSize, long? offset, long? maxReadBytes);
     }
 }
